Develop

+++ src/main/java/tech/gomes/reading/management/dto/book/response/CategoryFinishCountDTO.java
@@ -4,5 +4,5 @@ import lombok.Builder;
 
 @Builder
 public record CategoryFinishCountDTO(String categoryName,
-                                     double count) {
+                                     int count) {
 }
