+++ src/main/java/tech/gomes/reading/management/repository/projections/CategoryFinishCountProjection.java
@@ -2,5 +2,5 @@ package tech.gomes.reading.management.repository.projections;
 
 public interface CategoryFinishCountProjection {
     String getCategory();
-    double getCount();
+    Integer getCount();
 }
